scrapers/Edexcel/International/requirements.txt
# Edexcel International Qualifications Scraper Requirements

# Environment and configuration
python-dotenv>=1.0.0

# Database
supabase>=2.0.0

# HTTP requests
requests>=2.31.0

# PDF extraction (install both for best results)
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Optional: Better PDF handling
# pypdf>=3.17.0

